pyproject.toml
[project]
name = "rpoe-x"
version = "0.1.0"
description = "RPOE-X: Multi-Agent Rotary Parking Optimization Environment"
requires-python = ">=3.14"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "numpy",
    "python-dotenv",
    "openai",
    "matplotlib>=3.10.8",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "tasks"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
